backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
django-import-export==3.3.0
unfold==0.20.0
